pyproject.toml
[tool.poetry]
name = "python-project-boilerplate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
black = "^22.3.0"
pylint = "^2.13.4"
autoflake = "^1.4"
pydantic = "^1.9.0"
alembic = "^1.7.7"
SQLAlchemy = "^1.4.34"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
aiogram = "^2.19"
redis = "^4.2.1"
aioredis = "^2.0.1"
pandas = "^1.4.2"
openpyxl = "^3.0.9"
emoji = "^1.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
